GH-321 - PR comments
