+++ src/blockchain/src/blockchain.rs
@@ -13,13 +13,16 @@
    See the License for the specific language governing permissions and
    limitations under the License.
 */
-use super::block::*;
-use super::header::*;
+use std::collections::HashMap;
+use std::fmt;
+use std::fmt::{Debug, Display, Formatter};
+
 use libp2p::identity;
 use rand::Rng;
 use serde::{Deserialize, Serialize};
-use std::collections::HashMap;
-use std::fmt::{Display, Formatter};
+
+use super::block::*;
+use super::header::*;
 
 /// BlockchainId identifies the current chain
 #[derive(Serialize, Deserialize, Debug, Clone)]
@@ -44,7 +47,8 @@ pub enum SignatureAlgorithm {
 pub struct Config {
     pub blockchain_id: BlockchainId,
     pub hash_algorithm: HashAlgorithm,
-    pub hash_size: u32, //sizes of u8
+    pub hash_size: u32,
+    //sizes of u8
     pub signature_algorithm: SignatureAlgorithm,
     pub key_size: u32,
 }
@@ -107,7 +111,55 @@ pub struct Blockchain {
     pub blocks: Vec<Block>,
 }
 
+impl Debug for Blockchain {
+    fn fmt(&self, f: &mut Formatter<'_>) -> fmt::Result {
+        let Blockchain {
+            trans_observers: _,
+            genesis_block,
+            blocks,
+            block_observers: _,
+        } = self;
+
+        f.debug_struct("Blockchain")
+            .field("genesis_block", genesis_block)
+            .field("blocks", blocks)
+            .field("trans_observers", &self.trans_observers.len())
+            .field("block_observers", &self.block_observers.len())
+            .finish()
+    }
+}
+
 impl Blockchain {
+    pub fn submit_transaction<CallBack: 'static + FnOnce(Transaction)>(
+        &mut self,
+        trans: Transaction,
+        on_done: CallBack,
+    ) -> &mut Self {
+        self.trans_observers.insert(trans, Box::new(on_done));
+        self
+    }
+
+    pub fn notify_transaction_settled(&mut self, trans: Transaction) {
+        // if there were no observers, we don't care
+        if let Some(on_settled) = self.trans_observers.remove(&trans) {
+            on_settled(trans)
+        }
+    }
+
+    pub fn add_block_listener<CallBack: 'static + FnMut(Block)>(
+        &mut self,
+        on_block: CallBack,
+    ) -> &mut Self {
+        self.block_observers.push(Box::new(on_block));
+        self
+    }
+
+    pub fn notify_block_event(&mut self, block: Block) -> &mut Self {
+        self.block_observers
+            .iter_mut()
+            .for_each(|notify| notify(block.clone()));
+        self
+    }
     #[warn(dead_code)]
     pub fn new(keypair: &identity::ed25519::Keypair) -> Self {
         Self {
@@ -121,6 +173,7 @@ impl Blockchain {
     #[warn(dead_code)]
     pub fn add_block(&mut self, block: Block) {
         self.blocks.push(block);
+        self.notify_block_event(self.blocks.last().expect("block must exist").clone());
     }
 }
 
@@ -158,6 +211,9 @@ impl Display for Blockchain {
 
 #[cfg(test)]
 mod tests {
+    use std::cell::Cell;
+    use std::rc::Rc;
+
     use super::*;
 
     #[test]
@@ -199,6 +255,80 @@ mod tests {
         ));
         assert_eq!(true, chain.blocks.last().unwrap().verify());
         assert_eq!(2, chain.blocks.len());
+        Ok(())
+    }
+
+    #[test]
+    fn test_add_trans_listener() -> Result<(), String> {
+        let keypair = generate_ed25519();
+        let ed25519_keypair = match keypair {
+            identity::Keypair::Ed25519(v) => v,
+            identity::Keypair::Rsa(_) => todo!(),
+            identity::Keypair::Secp256k1(_) => todo!(),
+        };
+        let local_id = hash(&get_publickey_from_keypair(&ed25519_keypair).encode());
+        let mut chain = Blockchain::new(&ed25519_keypair);
+
+        let transaction = Transaction::new(
+            PartialTransaction::new(
+                TransactionType::Create,
+                local_id,
+                "some transaction".as_bytes().to_vec(),
+                rand::thread_rng().gen::<u128>(),
+            ),
+            &ed25519_keypair,
+        );
+        let called = Rc::new(Cell::new(false));
+        chain
+            .submit_transaction(transaction.clone(), {
+                let called = called.clone();
+                let transaction = transaction.clone();
+                move |t: Transaction| {
+                    assert_eq!(transaction, t);
+                    called.set(true)
+                }
+            })
+            .notify_transaction_settled(transaction);
+        assert!(called.get());
+        Ok(())
+    }
+
+    #[test]
+    fn test_add_block_listener() -> Result<(), String> {
+        let ed25519_keypair = match generate_ed25519() {
+            identity::Keypair::Ed25519(v) => v,
+            identity::Keypair::Rsa(_) => todo!(),
+            identity::Keypair::Secp256k1(_) => todo!(),
+        };
+        let local_id = hash(&get_publickey_from_keypair(&ed25519_keypair).encode());
+
+        let block_header = Header::new(PartialHeader::new(
+            hash(b""),
+            local_id,
+            hash(b""),
+            1,
+            rand::thread_rng().gen::<u128>(),
+        ));
+
+        let block = Block::new(
+            block_header,
+            Vec::new(),
+            &identity::ed25519::Keypair::generate(),
+        );
+        let mut chain = Blockchain::new(&ed25519_keypair);
+        let called = Rc::new(Cell::new(false));
+
+        chain
+            .add_block_listener({
+                let called = called.clone();
+                let block = block.clone();
+                move |b: Block| {
+                    assert_eq!(block, b);
+                    called.set(true);
+                }
+            })
+            .notify_block_event(block);
+        assert!(called.get()); // called is still false
         Ok(())
     }
 }
