+++ src/blockchain/src/lib.rs
@@ -18,5 +18,4 @@ pub mod block;
 pub mod blockchain;
 pub mod header;
 pub mod network;
-pub mod notifier;
 pub mod storage;
